feat(jwt): parse and validate
