+++ src/jwt/__fixtures__/wrong_public.pem
@@ -1,0 +1,9 @@
+-----BEGIN PUBLIC KEY-----
+MIIBITANBgkqhkiG9w0BAQEFAAOCAQ4AMIIBCQKCAQB/Wy3b6D6dy3x2N3nMJn42
+DLfyssUA+1usaAUCob7h82FoYVM9sm2dXhn8ddTTEWgnCDehMmbe7qkhAP3hgZ4p
+hp31Gb9XIkxeimmXZGi/8Rh2oA78CZwpRPewSb5E+pP9dRe4fJG/FSF5N4TFUv4O
+fMPPb1kcYgEorncC5RIlSgGiUvyYWlwuG5o4rWiIrP1A+rALiTOq97q0aqO6Cs/P
+aZ9fDCU7CVdDzvLVX/jcn1P0gZQpu3/3zO8IVpria30MA8jLtmRfQFkU++8NHm+P
+IsPJe4KqBO6unWEyDSIh1RoQVXZbFyE2Q2skPjPp55MVYyUOE6j6RszshhxFLSW9
+AgMBAAE=
+-----END PUBLIC KEY-----
