+++ src/jwt/const.js
@@ -1,0 +1,4 @@
+export const JWT_ALGORITHMS = {
+	RSA: 'RS512',
+	HMAC: 'HS512'
+};
